# api/openapi.yaml
openapi: 3.0.3
info:
  title: Flywheel API
  description: Local operator REST API for Flywheel projects, tickets, reviews, and workflows. No login or bearer token is required; the server binds to loopback and enforces local Host and same-origin browser requests. MCP authenticates separately with agent credentials.
  version: 0.1.0
servers:
  - url: http://localhost:8090
    description: Local
paths:
...
  /orgs:
    get:
      operationId: ListOrgs
      summary: List organizations for the local operator
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Org'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
    post:
      operationId: CreateOrg
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrgRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Org'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
  /orgs/{orgID}/projects:
    get:
      operationId: ListProjectsByOrg
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Filter by project status; default active.
          schema:
            type: string
            enum: [active, closed, all]
            default: active
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      operationId: CreateProject
      parameters:
        - name: orgID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StructuredError'
...
components:
  schemas:
...
    StructuredError:
      type: object
      required: [error, code, retriable]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [lease_expired, unauthorized, not_found, conflict, invalid_input, forbidden, internal, not_implemented, project_closed]
        retriable:
          type: boolean
...
    Org:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        created_at:
          type: string
          format: date-time
    CreateOrgRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        org_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          description: Short project description.
        repo_url:
          type: string
        default_branch:
          type: string
          description: Branch to checkout when closing a work stream; default "main".
        tech_stack:
          type: array
          items:
            type: string
        context_pack:
          type: object
        status:
          type: string
          enum: [active, closed]
          description: active (default) or closed; list endpoints default to active only.
        dispatch_enabled:
          type: boolean
          description: Whether the dispatcher picks up tickets for this project (default true).
        dispatch_config:
          $ref: '#/components/schemas/DispatchConfig'
        created_at:
          type: string
          format: date-time
...
    DispatchConfig:
      type: object
      properties:
        max_active_workers:
          type: integer
          minimum: 0
          description: Maximum concurrent active workers for this project. Set 0 or omit to inherit the server DISPATCH_MAX_WORKERS default.
        roles:
          type: array
          description: Custom dispatch roles. Each role has its own routing policy and inherits prompt/tool behavior from a built-in worker type.
          items:
            $ref: '#/components/schemas/DispatchWorkerRole'
        workers:
          type: array
          items:
            $ref: '#/components/schemas/DispatchWorkerProfile'
        policies:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/DispatchRolePolicy'
    DispatchWorkerRole:
      type: object
      properties:
        id:
          type: string
          description: Stable custom role key used in dispatch policies and ticket inputs.
        name:
          type: string
        description:
          type: string
          description: Additional instructions appended to workers launched for this custom role.
        base_type:
          type: string
          enum: [planner, executor, validator, deployer, investigator]
          description: Built-in worker behavior this custom role inherits for prompts and tool access.
    DispatchWorkerProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        runner:
          type: string
          enum: [cli, docker, openai-responses, openai-compatible]
        driver:
          type: string
          enum: [claude, generic, codex]
        cli_path:
          type: string
          description: Override binary path when using a CLI or Docker-backed worker.
        model:
          type: string
          description: Model override for API-native workers.
        reasoning_effort:
          type: string
          description: Reasoning effort override for API-native workers.
        api_base_url:
          type: string
          description: Base URL override for API-native workers.
        credential_env_var:
          type: string
          description: Environment variable that holds the provider credential for this worker profile.
        args:
          type: array
          items:
            type: string
        system_prompt:
          type: string
          description: Base prompt prepended for this worker
    DispatchRolePolicy:
      type: object
      properties:
        selection_mode:
          type: string
          enum: [ordered, any]
          description: ordered tries workers in the configured order; any rotates across the configured workers.
        worker_ids:
          type: array
          description: Worker profile IDs to use for this role. Leave empty to use all enabled workers.
          items:
            type: string
...
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        repo_url:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        template_id:
          type: string
          description: "Project template ID to seed from"
        work_stream_template_ids:
          type: array
          items:
            type: string
          description: "Which workstream templates to include. If omitted with template_id, all template workstreams are included."
...